ptyknot

Function ptyknot

pub fn ptyknot<F: FnOnce()>(
    action: F,
    pty: Option<&File>,
    plumbing: &[&Plumbing],
) -> Result<PtyKnot>

Start a child process running the given action, returning a PtyKnot for process information (currently just a process ID). When the the structure’s destructor is called, it will wait to reap the child process and panic if it has crashed or exited with non-zero status.

The optional pty argument, if supplied with the master side of a pseudoterminal as created by make_pty(), will cause the child to be set up with the slave side of that pseudoterminal as its controlling terminal. Otherwise, the child will be set up with no controlling terminal.

The action must not access the PTY or plumbing values passed to this function. Their child-side descriptors are closed or reassigned before the action runs.

Examples

use std::fs::OpenOptions;
use std::io::{Write, BufRead, BufReader};

fn slave() {
    let mut tty = OpenOptions::new()
                  .write(true)
                  .open("/dev/tty")
                  .expect("cannot open /dev/tty");
    tty.write("hello world\n".as_bytes())
       .expect("cannot write to /dev/tty");
    tty.flush().expect("cannot flush /dev/tty");
}

let mut pty = ptyknot::make_pty().expect("could not make pty");
let knot = ptyknot::ptyknot(slave, Some(&mut pty), &vec![])
           .expect("cannot create slave");
let mut tty = BufReader::new(&pty);
let mut message = String::new();
tty.read_line(&mut message)
   .expect("could not read message");
// This will wait for the child.
drop(knot);