ptyknot

Macro ptyknot

macro_rules! ptyknot {
    ($knot:ident,
     $slave:expr
     $(, @ $tty:ident)*
     $(, < $master_read:ident $read_fd:expr)*
     $(, > $master_write:ident $write_fd:expr)*) => { ... };
}

Provide a cleaner interface to ptyknot() et al by doing variable declaration and redeclaration. The first argument is the identifier for the resulting knot. The second argument is the child action, as with ptyknot(). The rest of the arguments are:

Zero or one pty redirections, consisting of @ followed by a pseudo-tty identifier.

Zero or more master read redirections, consisting of < followed by a master read identifier and a standard file descriptor expression.

Zero or more master write redirections, consisting of > followed by a master write identifier and a standard file descriptor expression.

The macro will let-declare the necessary handles, assemble them and pass them to ptyknot(), then redeclare the handles to allow the master to manipulate them.

#Example

use std::fs::OpenOptions;
use std::io::{Write, BufRead, BufReader};

fn slave() {
    let mut tty = OpenOptions::new()
                  .write(true)
                  .open("/dev/tty")
                  .expect("cannot open /dev/tty");
    tty.write("hello world\n".as_bytes())
       .expect("cannot write to /dev/tty");
    tty.flush().expect("cannot flush /dev/tty");
}

ptyknot!(knot, slave, @ pty);
let mut tty = BufReader::new(&pty);
let mut message = String::new();
tty.read_line(&mut message)
   .expect("could not read message");
// This will wait for the child.
drop(knot);