ptyknot

Module pty

Low-level pseudo-tty setup routines.

This module contains several pty setup functions and a “special” waitpid() implementation.

Much of this is code borrowed from http://github.com/stemjail/tty-rs.

Functions

dup2

Make the underlying file of dst refer to the underlying file of src. If dst is open, it will be closed first. See dup2(2) in the UNIX manual pages for details.

grantpt

Change the mode and owner of the slave pty associated with a given open master. See grantpt(3) in the UNIX manual pages for details.

pipe

Make a new pipe. The 0-th side of the resulting array is the read side. The 1-th side is the write side. See pipe(2) in the UNIX manual pages for details.

ptsname

Returns the name of the slave pty associated with the given open master. See unlockpt(3) in the UNIX manual pages for details.

unlockpt

“Unlocks” the slave pty associated with the give nopen master. See unlockpt(3) in the UNIX manual pages for details.

waitpid

Blocking wait until process completes. Returns the child process status. See waitpid(3) in the UNIX manual pages for details.