Struct ptyknot::Plumbing

pub struct Plumbing { /* fields omitted */ }

Information needed during the pipe plumbing process.

Methods

impl Plumbing

fn new(direction: PipeDirection, slave_target: RawFd) -> Result<Plumbing>

Create a new pipe running in the specified direction, and remember the file descriptor of the given file. This will later allow the slave to attach slave_target to the other end of the pipe.

fn plumb_slave(&self) -> Result<()>

Implement the slave side of the plumbing by ensuring that the slave end of the pipe is attached to the previously-supplied file descriptor.

fn get_master(self) -> Result<File>

Extract the master side of the pipe for use by the parent.