Crate ptyknot

Child process and pty support.

Start a child process running a specified action. The process may have a new pseudo-tty as its controlling terminal, and may have pipes to the master for some of its file descriptors. The caller receives handles for all of this, along with the process ID of the child. The caller can then later wait for the child to exit by dropping its last reference.

Modules

pty

Low-level pseudo-tty setup routines.

Macros

ptyknot

Provide a cleaner interface to ptyknot() et al by doing variable declaration and redeclaration. The first argument is the identifier for the resulting knot. The second argument is the child action, as with ptyknot(). The rest of the arguments are:

Structs

Plumbing

Information needed during the pipe plumbing process.

PtyKnot

Parent information about the child process.

Enums

PipeDirection

Which direction a pipe runs.

Functions

make_pty

Return the master side of a ready-to-operate pseudo-terminal.

pty_stderr

Stderr for pty-side interface.

pty_stdin

Stdin for pty-side interface.

pty_stdout

Stdout for pty-side interface.

ptyknot

Start a child process running the given action, returning a PtyKnot for process information (currently just a process ID). When the the structure’s destructor is called, it will wait to reap the child process and panic if it has crashed or exited with non-zero status.